requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.4
django-debug-toolbar==6.3.0
iniconfig==2.3.0
kombu==5.6.2
packaging==26.1
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
